Build fix for GMP.

+++ mplapack/reference/Rggevx.cpp
@@ -422,11 +422,11 @@ void Rggevx(const char *balanc, const char *jobvl, const char *jobvr, const char
             temp = zero;
             if (alphai[jc - 1] == zero) {
                 for (jr = 1; jr <= n; jr = jr + 1) {
-                    temp = max(temp, abs(vl[(jr - 1) + (jc - 1) * ldvl]));
+                    temp = max(temp, REAL(abs(vl[(jr - 1) + (jc - 1) * ldvl])));
                 }
             } else {
                 for (jr = 1; jr <= n; jr = jr + 1) {
-                    temp = max(temp, abs(vl[(jr - 1) + (jc - 1) * ldvl]) + abs(vl[(jr - 1) + ((jc + 1) - 1) * ldvl]));
+                    temp = max(temp, REAL(abs(vl[(jr - 1) + (jc - 1) * ldvl]) + abs(vl[(jr - 1) + ((jc + 1) - 1) * ldvl])));
                 }
             }
             if (temp < smlnum) {
@@ -455,11 +455,11 @@ void Rggevx(const char *balanc, const char *jobvl, const char *jobvr, const char
             temp = zero;
             if (alphai[jc - 1] == zero) {
                 for (jr = 1; jr <= n; jr = jr + 1) {
-                    temp = max(temp, abs(vr[(jr - 1) + (jc - 1) * ldvr]));
+                    temp = max(temp, REAL(abs(vr[(jr - 1) + (jc - 1) * ldvr])));
                 }
             } else {
                 for (jr = 1; jr <= n; jr = jr + 1) {
-                    temp = max(temp, abs(vr[(jr - 1) + (jc - 1) * ldvr]) + abs(vr[(jr - 1) + ((jc + 1) - 1) * ldvr]));
+                    temp = max(temp, REAL(abs(vr[(jr - 1) + (jc - 1) * ldvr]) + abs(vr[(jr - 1) + ((jc + 1) - 1) * ldvr])));
                 }
             }
             if (temp < smlnum) {
