+++ mplapack/reference/Rtprfs.cpp
@@ -247,9 +247,9 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
         s = zero;
         for (i = 1; i <= n; i = i + 1) {
             if (work[i - 1] > safe2) {
-                s = max(s, abs(work[(n + i) - 1]) / work[i - 1]);
+                s = max(s, REAL(abs(work[(n + i) - 1]) / work[i - 1]));
             } else {
-                s = max(s, (abs(work[(n + i) - 1]) + safe1) / (work[i - 1] + safe1));
+                s = max(s, REAL((abs(work[(n + i) - 1]) + safe1) / (work[i - 1] + safe1)));
             }
         }
         berr[j - 1] = s;
@@ -312,7 +312,7 @@ void Rtprfs(const char *uplo, const char *trans, const char *diag, INTEGER const
         //
         lstres = zero;
         for (i = 1; i <= n; i = i + 1) {
-            lstres = max(lstres, abs(x[(i - 1) + (j - 1) * ldx]));
+            lstres = max(lstres, REAL(abs(x[(i - 1) + (j - 1) * ldx])));
         }
         if (lstres != zero) {
             ferr[j - 1] = ferr[j - 1] / lstres;
